Run and publish unit tests
