--- invoice-app/RandomNames/models/RandomPerson.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace invoice_app.RandomNames.models
-{
-    public class RandomPerson
-    {
-        public string Firstname { get; set; }
-        public string Middlename { get; set; }
-        public string Lastname { get; set; }
-    }
-}
