+++ invoice-app/invoice-app.csproj
@@ -24,10 +24,8 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.1" />
-    <PackageReference Include="Microsoft.Office.Interop.Excel" Version="15.0.4795.1000" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.0" />
     <PackageReference Include="PdfSharp" Version="1.50.5147" />
-    <PackageReference Include="RandomNameGeneratorLibrary" Version="1.2.2" />
     <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
   </ItemGroup>
 
@@ -39,8 +37,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Basic\dto\" />
-    <Folder Include="Basic\services\impl\" />
     <Folder Include="Migrations\" />
     <Folder Include="Models\" />
   </ItemGroup>
